Fixed decommisisons and goods arrival

+++ backend/app/db/models.py
@@ -162,7 +162,7 @@ class AccessoriesStorage(Base):
     __tablename__ = "accessories_storage"
 
     article = Column(Integer, ForeignKey("accessories.article"), primary_key=True)
-    amount = Column(Numeric(10, 3))
+    amount = Column(Integer)
 
 
 class AccessoriesStorageWithAccessory(AccessoriesStorage):
@@ -209,7 +209,7 @@ class AccessoryChanges(Base):
     accessory_article = Column(Integer, ForeignKey("accessories.article"))
     is_income = Column(Boolean, default=True)
     timestamp = Column(DateTime, server_default=func.now())
-    amount = Column(Numeric(10, 3))
+    amount = Column(Integer)
 
 
 class ClothChanges(Base):
